+++ src/components/HarvestingConfiguration/VendorInfo/VendorInfoForm.js
@@ -27,12 +27,12 @@ const VendorInfoForm = ({
         <Col xs={4}>
           <Field
             component={Select}
+            data={!disabled && isRequired ? 1 : 0}
             dataOptions={harvesterImpls}
-            defaultValue="cs50"
+            defaultValue="cs51"
             disabled={disabled}
             fullWidth
             id="addudp_servicetype"
-            data={!disabled && isRequired ? 1 : 0}
             label={<FormattedMessage id="ui-erm-usage.vendorInfo.serviceType" />}
             name="harvestingConfig.sushiConfig.serviceType"
             required={!disabled && isRequired}